refactor: Move validation code out of StatusConfiguration
To StatusValidator

Part of addressing #1535
